mudifica os diretorios do projeto
