+++ backend/src/main/java/com/justice/justiceforall/service/casesservice/CreateCaseValidator.java
@@ -1,10 +1,17 @@
-package com.justice.justiceforall.service.casesservice;
+package com.justice.justiceforall.service.casesservice.casevalidator;
 
 import com.justice.justiceforall.dto.casesdto.CreateCaseCommand;
 
+import lombok.experimental.UtilityClass;
+
+@UtilityClass
 public class CreateCaseValidator {
-
-    public static void validateNewCase(CreateCaseCommand createCaseCommand) {
+	void validateNewCase(CreateCaseCommand createCaseCommand) {
+        var fileFormatValidator = new FileFormatValidator();
+        var mustHaveFields = new MustHaveFields();
+        
+        fileFormatValidator.setNext(mustHaveFields);
+     
+        fileFormatValidator.validate(createCaseCommand);
     }
-
 }
